m3-capstone-project/pyproject.toml
[tool.poetry]
name = "agent-resume"
version = "0.1.0"
description = "Multi-agent HR RAG system with supervisor/worker architecture"
authors = ["you <you@example.com>"]
readme = "README.md"
packages = [{ include = "agent_humancapital", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
streamlit = "^1.36.0"
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"
pandas = "^2.2.2"
openpyxl = "^3.1.5"

qdrant-client = "^1.9.1"
langchain = "^0.2.12"
langchain-core = "^0.2.28"
langchain-openai = "^0.1.21"
langchain-qdrant = "^0.1.4"
tiktoken = "^0.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
